requirements.txt
asgiref==3.2.3
Django==3.0.3
django-filter==2.2.0
django-resized==0.3.9
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-rest-swagger==2.1.2
django-cors-headers==3.2.
drf-yasg==1.17.1
djoser==2.0.3
fcm-django==0.3.4
pyfcm==1.4.7
Markdown==3.2.1
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
